@@ -1,6 +1,6 @@
 {
   "name": "react-native-gifted-chat",
-  "version": "0.0.10",
+  "version": "0.1.3",
   "description": "The most complete chat UI for React Native",
   "main": "index.js",
   "repository": {
@@ -23,19 +23,27 @@
     "url": "https://github.com/FaridSafi/react-native-gifted-chat/issues"
   },
   "homepage": "https://github.com/FaridSafi/react-native-gifted-chat#readme",
+  "scripts": {
+    "lint": "eslint . --ext .js,.jsx"
+  },
   "devDependencies": {
-    "eslint": "^3.2.2",
-    "eslint-plugin-react": "^6.0.0",
-    "eslint-plugin-react-native": "^2.0.0"
+    "eslint": "~3.12.2",
+    "eslint-config-airbnb": "~13.0.0",
+    "eslint-config-cooperka": "~0.2.3",
+    "eslint-plugin-import": "~2.2.0",
+    "eslint-plugin-jsx-a11y": "~2.2.3",
+    "eslint-plugin-react": "~6.8.0",
+    "eslint-plugin-react-native": "~2.2.1"
   },
   "dependencies": {
-    "@exponent/react-native-action-sheet": "git+https://github.com/FaridSafi/react-native-action-sheet.git",
-    "md5": "^2.1.0",
-    "moment": "^2.13.0",
-    "react-native-communications": "^2.1.0",
-    "react-native-dismiss-keyboard": "^1.0.0",
-    "react-native-invertible-scroll-view": "^1.0.0",
-    "react-native-parsed-text": "^0.0.15",
-    "shallowequal": "^0.2.2"
+    "@exponent/react-native-action-sheet": "0.3.0",
+    "md5": "2.2.1",
+    "moment": "2.17.1",
+    "react-native-communications": "2.1.1",
+    "react-native-invertible-scroll-view": "1.0.0",
+    "react-native-lightbox": "oblador/react-native-lightbox#c84a8543d4511fe6a44c3d7820747c9c1bddd875",
+    "react-native-parsed-text": "0.0.16",
+    "shallowequal": "0.2.2",
+    "uuid": "3.0.1"
   }
 }
