@@ -1,11 +1,7 @@
-import React from 'react';
-import {
-  Image,
-  StyleSheet,
-  View,
-} from 'react-native';
-
-import GiftedAvatar from './GiftedAvatar';
+import React from "react";
+import {Image, StyleSheet, View} from "react-native";
+import GiftedAvatar from "./GiftedAvatar";
+import {isSameUser, isSameDay, warnDeprecated} from "./utils";
 
 export default class Avatar extends React.Component {
   renderAvatar() {
@@ -22,7 +18,11 @@ export default class Avatar extends React.Component {
   }
 
   render() {
-    if (this.props.isSameUser(this.props.currentMessage, this.props.nextMessage) && this.props.isSameDay(this.props.currentMessage, this.props.nextMessage)) {
+    const renderAvatarOnTop = this.props.renderAvatarOnTop;
+    const messageToCompare = renderAvatarOnTop ? this.props.previousMessage : this.props.nextMessage;
+    const computedStyle = renderAvatarOnTop ? "onTop" : "onBottom"
+
+    if (isSameUser(this.props.currentMessage, messageToCompare) && isSameDay(this.props.currentMessage, messageToCompare)) {
       return (
         <View style={[styles[this.props.position].container, this.props.containerStyle[this.props.position]]}>
           <GiftedAvatar
@@ -32,7 +32,8 @@ export default class Avatar extends React.Component {
       );
     }
     return (
-      <View style={[styles[this.props.position].container, this.props.containerStyle[this.props.position]]}>
+      <View
+        style={[styles[this.props.position].container, styles[this.props.position][computedStyle], this.props.containerStyle[this.props.position]]}>
         {this.renderAvatar()}
       </View>
     );
@@ -42,8 +43,12 @@ export default class Avatar extends React.Component {
 const styles = {
   left: StyleSheet.create({
     container: {
-      marginRight: 8,
+      marginRight: 8
     },
+    onTop: {
+      alignSelf: "flex-start"
+    },
+    onBottom: {},
     image: {
       height: 36,
       width: 36,
@@ -54,6 +59,10 @@ const styles = {
     container: {
       marginLeft: 8,
     },
+    onTop: {
+      alignSelf: "flex-start"
+    },
+    onBottom: {},
     image: {
       height: 36,
       width: 36,
@@ -63,8 +72,7 @@ const styles = {
 };
 
 Avatar.defaultProps = {
-  isSameDay: () => {},
-  isSameUser: () => {},
+  renderAvatarOnTop: false,
   position: 'left',
   currentMessage: {
     user: null,
@@ -72,14 +80,25 @@ Avatar.defaultProps = {
   nextMessage: {},
   containerStyle: {},
   imageStyle: {},
+  //TODO: remove in next major release
+  isSameDay: warnDeprecated(isSameDay),
+  isSameUser: warnDeprecated(isSameUser)
 };
 
 Avatar.propTypes = {
-  isSameDay: React.PropTypes.func,
-  isSameUser: React.PropTypes.func,
+  renderAvatarOnTop: React.PropTypes.bool,
   position: React.PropTypes.oneOf(['left', 'right']),
   currentMessage: React.PropTypes.object,
   nextMessage: React.PropTypes.object,
-  containerStyle: View.propTypes.style,
-  imageStyle: React.PropTypes.oneOfType([View.propTypes.style, Image.propTypes.style]),
+  containerStyle: React.PropTypes.shape({
+    left: View.propTypes.style,
+    right: View.propTypes.style,
+  }),
+  imageStyle: React.PropTypes.shape({
+    left: View.propTypes.style,
+    right: View.propTypes.style,
+  }),
+  //TODO: remove in next major release
+  isSameDay: React.PropTypes.func,
+  isSameUser: React.PropTypes.func
 };
