@@ -6,18 +6,33 @@ import {
 } from 'react-native';
 
 export default class Composer extends React.Component {
+  onChange(e) {
+    const contentSize = e.nativeEvent.contentSize;
+    if (!this.contentSize) {
+      this.contentSize = contentSize;
+      this.props.onInputSizeChanged(this.contentSize);
+    } else if (this.contentSize.width !== contentSize.width || this.contentSize.height !== contentSize.height) {
+      this.contentSize = contentSize;
+      this.props.onInputSizeChanged(this.contentSize);
+    }
+  }
+
+  onChangeText(text) {
+    this.props.onTextChanged(text);
+  }
+
   render() {
     return (
       <TextInput
         placeholder={this.props.placeholder}
         placeholderTextColor={this.props.placeholderTextColor}
         multiline={this.props.multiline}
-        onChange={(e) => {
-          this.props.onChange(e);
-        }}
-        style={[styles.textInput, this.props.textInputStyle, {
-          height: this.props.composerHeight,
-        }]}
+
+        onChange={(e) => this.onChange(e)}
+        onChangeText={text => this.onChangeText(text)}
+
+        style={[styles.textInput, this.props.textInputStyle, {height: this.props.composerHeight}]}
+
         value={this.props.text}
         accessibilityLabel={this.props.text || this.props.placeholder}
         enablesReturnKeyAutomatically={true}
@@ -46,7 +61,8 @@ const styles = StyleSheet.create({
 });
 
 Composer.defaultProps = {
-  onChange: () => {},
+  onChange: () => {
+  },
   composerHeight: Platform.select({
     ios: 33,
     android: 41,
@@ -57,6 +73,10 @@ Composer.defaultProps = {
   textInputProps: null,
   multiline: true,
   textInputStyle: {},
+  onTextChanged: () => {
+  },
+  onInputSizeChanged: () => {
+  },
 };
 
 Composer.propTypes = {
@@ -66,6 +86,8 @@ Composer.propTypes = {
   placeholder: React.PropTypes.string,
   placeholderTextColor: React.PropTypes.string,
   textInputProps: React.PropTypes.object,
+  onTextChanged: React.PropTypes.func,
+  onInputSizeChanged: React.PropTypes.func,
   multiline: React.PropTypes.bool,
   textInputStyle: TextInput.propTypes.style,
 };
