@@ -140,10 +140,11 @@ export default class MessageContainer extends React.Component {
       <View ref='container' style={{flex:1}}>
         <ListView
           enableEmptySections={true}
-          keyboardShouldPersistTaps={true}
           automaticallyAdjustContentInsets={false}
           initialListSize={20}
           pageSize={20}
+
+          {...this.props.listViewProps}
 
           dataSource={this.state.dataSource}
 
@@ -162,6 +163,7 @@ MessageContainer.defaultProps = {
   user: {},
   renderFooter: null,
   renderMessage: null,
+  listViewProps: {},
   onLoadEarlier: () => {
   },
 };
@@ -172,4 +174,5 @@ MessageContainer.propTypes = {
   renderFooter: React.PropTypes.func,
   renderMessage: React.PropTypes.func,
   onLoadEarlier: React.PropTypes.func,
+  listViewProps: React.PropTypes.object,
 };
